BST-22 About Dialog tab issue fixed
Fixed the issue with tabs with long text get cut off on the edges. Issue only happened on certain iMacs, I suspect the reason for this has to do with high-res monitors and as long as the tab with is below 210px the issue does not occur. After, I found the problem css line with trial and error. Removing the font setting does not seem to affect the fonts in the app as they're inherited.