Fix RBF fee estimation
The RBF algorithm is currently incorrect, and forces the user to pay too much in fees. Follow the RBF rules more closely so that users aren't forced to pay more than they absolutely must.

Note that, due to the nature of how the RBF minimum fee is increased, it's impossible to determine the absolute minimum until the TX is signed. Therefore, the terminal merely provides an estimate of the floor, albeit one that be off at most by 2 satoshis (but usually only 1 if it is off) per input or output. Unless a user has hundreds or thousands of inputs or outputs, the excess fees will amount to little more than dust.