Workaround for an Armory wallet bug
Due to a bug in Armory that hasn't been fixed yet in 0.97 (the 0.96.5 fix isn't applicable as-is), attempts to spend P2SH-P2WPKH inputs will lead to virtual size estimations that are incorrect. This will cause RBF issues. Insert a workaround that increases the virtual size by an amount observed to keep the virtual size close to the final, signed virtual size.