BST-1244 - Correct ZC wallet balances
Right now, when zero conf TXs come in, the spendable and unconfirmed balances are incorrect. Correct them.

Note that this code assumes that ZC UTXOs are spendable, as this is the current state of the code (and is allowed in Bitcoin). If this is unintentional, this behavior should be fixed in a separate commit.