Set macOS minimum version to 10.12
Right now, the terminal seems to set the macOS build minimum version to the minor version of the build system (e.g., a system on 10.13.6 will set the minimum to 10.13.0). Set the minimum to 10.12, which is the minimum required by Qt 5.12 (the first version to properly support macOS's "Dark Mode").