Partial fix for address details on ZC TX arrival
When a ZC TX arrives, the wallet & address detail widgets don't update an address's # of Txs or its balance. Update the balance.

Note that, for now, the # of TXs isn't updated. Armory doesn't update the number on the DB until the first conf. (The balance on the DB is updated when the ZC is processed.) Hacking this on the terminal side isn't trivial. This needs to be discussed further.

In addition, some minor logging cleanup was done, and a few comments were added.