Update Armory code (commit 2a696d4)
Update codebase to a later Armory revision with BIP 150/151 included (but not activated). The new code compiles on Linux (Ubuntu 18.10) and macOS but isn't used for anything yet. (A future commit will include changes required to active BIP 150/151 support.) Note that the libbtc changes from the commits aren't included.

They'll need to be pulled into a separate libbtc repo.