Fix TX virtual size calculation
When creating a TX, we need an accurate estimate of the TX virtual size. Properly initialize the relevant data, and get the resultant weight. Also fix some misleading UI statements ("weight" --> "virtual size").

Note that this TX has a portion that can be changed later if a proposed change is upstreamed in Armory. It works fine as-is but should be swapped out eventually to remove unnecessarily redundant work.