Add null character when using libbtc's base58 encode method
Armory's base58 encoding method added a null character to the string to be encoded. The libbtc method assumes a null character is already present. Add the character, and be more consistent in passing in BinaryData objects instead of string objects.

Note that this is a minor hack of sorts. A PR will be developed for Armory that will check incoming data for a null character, and add it if necessary. At that point, it won't matter if the incoming string has a null character, and the append() calls in this PR can be removed.