Fix P2SH virtual size (+ Armory upgrade)
- P2SH in Armory uses a nesting strategy in order to fully reveal an AddressEntry object's input and witness data size. Use the nesting in order to fully resolve a P2SH input's size.
- Upgrade Armory (commit 9e4e8c1e) to include the parallel functionality. (The terminal can't access the proper code, so it basically mirrors the proper code.)

Note that Armory's TX size calculations are a little conservative and will be a bit high. For example, a TX with a P2SH-P2WPKH input and two P2SH-P2WPKH outputs will return a virtual size of 169. The actual virtual size picked up by Core will be 166 or 167, depending on the signature size. There is room for improvement in Armory but the code as-is will ensure that users can safely use a fee of 1 sat/byte.