Fix MAX button fee calculation
When calculating the maximum accepted fee for a TX, the data was improperly initialized and caused Armory to calculate incorrect amounts. Properly initialize the data before asking Armory to calculate the maximum acceptable amount to send.

Note that the maximum fee will occasionally be a few satoshis higher than the true minimum that will be accepted on the Bitcoin network. This is due to how Armory calculates the fee, and it is acceptable. It can be revisited if a single satoshi ever becomes valuable enough to notice.