Fix if/else conditions when searching for internal addresses
When checking to see if an internal address has <6 confs, the if/else checks silently failed. Fix them.

This commit is to be combined with a more comprehensive commit that will fix the balance issues and ensure that internal address balances are properly adjusted. This commit was required in order to properly troubleshoot the overarching issue.