Update Armory (commit a2d568c)
The purpose of the commit is to add a message to the beginning of a WebSocket server connection where the server sends its own public key. This will set up the use case where the terminal won't need a public key for every ArmoryDB server. It'll only need a root public key and a derivation proof that can be used to confirm that the server's public key is acceptable.

This commit is required to connect to the latest ArmoryDB on the server.