Upgrade Qt to 5.12.2 (#247)
* [DO NOT CHECK IN YET] Upgrade Qt to 5.12

Among other things, Qt 5.12 adds support for Dark Mode in macOS 10.14. This is necessary in order for apps to not look like garbage on Macs.

* [DO NOT CHECK IN YET] Upgrade Qt to 5.12

Among other things, Qt 5.12 adds support for Dark Mode in macOS 10.14. This is necessary in order for apps to not look like garbage on Macs.

* [DO NOT CHECK IN YET] Upgrade Qt to 5.12

Among other things, Qt 5.12 adds support for Dark Mode in macOS 10.14. This is necessary in order for apps to not look like garbage on Macs.

* Upgrade Qt to 5.12.1

Among other things, Qt 5.12 is an LTR (5.11 isn't) that adds support for Dark Mode in macOS 10.14. This is necessary in order for apps to not look like garbage on Macs.

* Set Qt to 5.12.2

The third time's the charm? The OpenSSL linking issues in Windows should be gone now with 5.12.2.