Fix TX explorer crash
If Armory hasn't fully initialized yet, the TX explorer can't connect to Armory and also can't get a copy of the logger object. Writes to the logger in such a case cause an instant crash. Write to the logger only when the logger object has been initialized.

Also perform some minor formatting cleanup.