Allow Armory to connect even when no wallets are present (#698)
* Allow Armory to connect even when no wallets are present

- Users may wish to use the explorer without having a wallet. Allow Armory to connect even if no wallets are present.
- Logger cleanup.

* Armory connection fixup

- Remove unnecessary goOnline() call.

* Review fixup

Getting Armory online requires wallets to be synched (assuming any exist) and for a particular Armory connection state to have been hit. Make sure all conditions are met before attempting to go online.

* Add more conditions for wallet-less Armory connections

After feedback, it was determined that the online condition needed to add a check for any wallets being present in the wallet manager. Armory should go online only if no wallets are present, including the settlement wallet (although the latter condition is commented out for now due to a bug that will be fixed elsewhere).