Temporarily set ZMQ BIP15X rekey timer to one week
While a couple of bugs are hashed out, the rekey timer is being set to 604,800 seconds (1 week) in order to (hopefully) prevent the heartbeat from tripping up the chat server and any other services using this particular connection type. They'll still need to be rebooted, only after a week instead of ten minutes.