Move function body
We have qt types registration hidden inside module. And there are only few functions that will be a part of
generated code. So if none of them is linked to app, static intialization code will not be linked/executed.
So just moving few function definitions related to MD provider to get common types registered in Qt.