Refactor the ZMQ BIP connection key handling (#203)
*  Refactor the ZMQ BIP connection key handling

When we have a local signer, the signer will be invoked with the terminal's BIP 150 ID key, and the signer will generate a cookie with its ID key. This helps ensure that only the one given key will be accepted, thereby preventing attacks.

* Add required common submodule position

* Update common

* Update common one last time (for real!)

* Add proper support for the signer connection

As written, the data connections for the signer don't function properly with the required encryption. Add the required terminal code.

* Set common to new required position