Fix signer option parsing crash
DESCRIPTION:
When the signer binary is started with unsupported command line options, the signer crashes. Add a try/catch to ensure that the binary can continue.

TESTING:
On macOS, running blocksettle_signer with unsupported options (e.g., --i_do_not_work) caused the binary to crash and not start. With the fix, the binary starts, with a warning to the user that not all options may be respected. (The option parsing code, cxxopts, is limited in terms of error recovery capabilities.)