Testing connection close detection
So far, it seems to work. OnPeerDisconnect() is called each time.
SignerAdapterListener() seems to set to get this notification (in
no simple terms) and then sends a notification of its own about it.
Didn't track it further down. I guess a problem could be that this
notification carries the IP address, _not_ the Client ID and that is
not handled properly.