Add static ID key for remote signer in terminal/signer connection (terminal)
DESCRIPTION:
In order for the remote signer to do two-way verification with the terminal, the signer must have a static ID key. Refactor the code to allow for a new wallet key file (remote_signer.peers) that can be used for two-way verification in BIP 150. In addition, a related bug (always generate an ID key file, whether or not a local signer is used) was fixed.

TESTING:
Confirmed that the remote signer had a stable ID key after the code was written. Also confirmed that the local signer continues to use ephemeral keys. Connections were properly established in both cases. Testing was done on macOS.