Fix detection for client that never sent a heartbeat
Since such client would never insert a 'time of heartbeat' in the map,
it will never be checked against a time, and thus will never be detected
as disconnected. So we just act as-if the connection is the initial heartbeat
and thus map will always have entries for each client.