OTC Chat bug fixed and code refactoring
Bug fixed:
When user receive tray notifications for new messages and then click on notification, user should be moved to this chat. The previous code worked incorrect in the function to move user to required chat.

"Select first room" was renamed to "Select Global room".

Added new function "setCurrentUserChat" to ChatClientUserView and old code was rewritten to use this function.