OTC Chat: additional fix to set message was read if current room is selected (#893)
* Additional fix to set message was read if current room is selected

There was a bug when sometimes blue dot notification didn't disappear after message was read. PR #867 didn't fix that right, so the code of #867 was removed and added new fix which makes a check is a room with unread message selected.

* Code fixes

Previous code in ChatClientUsersViewItemDelegate was wrong. New code marks all messages in selected room as read in ChatMessagesTextEdit.