Clarify a log statement (#901)
* Clarify a log statement

DESCRIPTION:
A previous PR (#852) attempted to clarify a log statement in the ZMQ BIP15X server connection code. A revision of the log statement introduced a compile error, causing the statement to be reverted. Re-introduce the statement, with initial feedback included.

TESTING:
Confirmed that the code compiled and ran on Linux.

* const fixup