Fixed: blue dot of unread message don't disappear for room chats
The bug is related to OTC Chat.
Blue dot of unread message don't disappear for room chats even if all messages were read. It happens only if user which is not in contact list send a message to room chat. If user which sent message to room chat is in contact list, it works fine. Now it fixed.