Terminal - ui fixes (#1062)
* 1.
Settings
a.
Network – Armory Server – Manage
b.
Signer – Remote – manage
c.
When we’re inside these manage sections and highlight a row, this row
should be autoselected as the choice once we close the window. We
shouldn’t have to press select for the selection to “stick” once we go
back to the screen before Manage.
d.
Also, if we create a new connection, this should be selected by default.
2.
Settings - Signer
a.
Change so that we show a view like Remote at all times where we have
manage below. The “Local Headless” connection is a setting in the manage
section which is there by default and can’t be deleted. There should be
a second connection by default, the “Local GUI Mode” which is the
127.0.0.1 selection which we use if the local signer is run in GUI mode
and to which the terminal can auto-shift.
b.
Show Client side key at all times regardless of connection
c.
Have auto-signing grey if we are in remote mode

* Cleanup