Detect Celer disconnects
Sometimes Celer server detects disconnects if terminal do not send heartbeats for some time (for example under debugger) while proxy connection is still works.
Celer connection detected as live but no RFQ are sent or received.
This could be detected because Celer will stop sending heartbeats but we don't do this yet. This commit adds this.