add breakpad crash handler for Windows
Integrate and modify code from this demo:

https://github.com/KandaoVR/qt-breakpad

Use the latest version of breakpad with the cmake file from the vcpkg
port.

To enable the crash handler, pass the cmake option to `generate.py`,
e.g.:

```bash
python ./generate.py -cmake-flags '-DENABLE_CRASH_HANDLER=ON' release
```

It will then be activated for both the terminal and the signer.

When `OFF` breakpad will still be in effect to launch the systray
cleanup code, but the crash handler will not display.

I am working on the report receiver web app here:

https://github.com/rkitover/qt-breakpad-web

TODO:

- see if mac and linux can work

Signed-off-by: Rafael Kitover <rkitover@gmail.com>